Backend/requirements.txt
futu-api
psycopg2-binary
sqlalchemy
pandas
python-dotenv
apscheduler